requirements.txt
torch==2.5.1
torchvision
transformers==4.51.3
accelerate==1.2.1
sentencepiece==0.2.0
pytorch-extension
opencv-python==4.8.0.74
datasets
dataclasses
DateTime
typing
pillow
trl
peft
wandb
packaging
deepspeed
pandas
bitsandbytes
qwen_vl_utils
math_verify
peft